WorkerService: move to /services packages
