+++ app/src/main/AndroidManifest.xml
@@ -46,7 +46,7 @@
 
         <service android:name=".services.GcmIntentService" />
 
-        <service android:name=".worker.WorkerService" />
+        <service android:name=".services.WorkerService" />
 
         <receiver
             android:name=".services.GcmBroadcastReceiver"
