+++ app/src/main/java/com/nbusy/app/services/DeviceBootBroadcastReceiver.java
@@ -4,8 +4,6 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
-import com.nbusy.app.worker.WorkerService;
-
 public class DeviceBootBroadcastReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
