+++ app/src/main/java/com/nbusy/app/services/WorkerService.java
@@ -1,13 +1,13 @@
-package com.nbusy.app.worker;
+package com.nbusy.app.services;
 
 import android.app.Service;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Binder;
 import android.os.IBinder;
 import android.util.Log;
 
-import com.nbusy.app.services.DeviceBootBroadcastReceiver;
+import com.nbusy.app.worker.Worker;
+import com.nbusy.app.worker.WorkerSingleton;
 
 import java.util.Objects;
